+++ rollup.config.js
@@ -1,8 +1,8 @@
 import typescript from 'rollup-plugin-typescript2';
-import { terser } from "rollup-plugin-terser";
-import json from 'rollup-plugin-json';
-import commonjs from 'rollup-plugin-commonjs'
-import resolve from 'rollup-plugin-node-resolve';
+import terser from '@rollup/plugin-terser';
+import json from '@rollup/plugin-json';
+import commonjs from '@rollup/plugin-commonjs';
+import {nodeResolve} from '@rollup/plugin-node-resolve';
 import builtins from 'builtin-modules'
 import pkg from './package.json';
 
@@ -20,19 +20,16 @@ export default {
     terser(),
     json(),
     commonjs(),
-    resolve({
+    nodeResolve({
       mainFields: ['module', 'main'],
       // preferBuiltins: false
     })
   ],
   output: [
-    // ES module (for bundlers) build.
     {
       format: 'esm',
       file: pkg.module
-    },
-    // CommonJS (for Node) build.
-    {
+    }, {
       format: 'cjs',
       file: pkg.main
     }