+++ src/Browser.ts
@@ -6,6 +6,10 @@ import Bowser from 'bowser';
 export default class {
   /**
    * Analyze browser information from UA.
+   *
+   * @static
+   * @param {string} ua User agent string.
+   * @return {{platform: string, osName: string, osVersion: number|null, browserName: string}} Analysis Results.
    */
   public static parse(ua: string): {platform: string, osName: string, osVersion: number|null, browserName: string} {
     const parser = Bowser.getParser(ua);