+++ src/File.ts
@@ -9,7 +9,13 @@ const fse = require('fs-extra');
 
 export default class File {
   /**
-   * Get file name.
+   * Get the file name from the path.
+   *
+   * @static
+   * @param {string} filePath File Path.
+   * @param {boolean} withExtension True if you want to include the extension, false if you don't. Default is false.
+   * @return {string} File Name.
+   * @memberof File
    */
   public static basename(filePath: string, withExtension: boolean = false): string {
     return path.basename(filePath, withExtension ? undefined : path.extname(filePath));
@@ -17,15 +23,24 @@ export default class File {
 
   /**
    * Change permissions.
+   *
+   * @static
+   * @param {string} path File or directory path.
+   * @param {number} permission Permissions. Default is 0o755.
+   * @return {File}
+   * @memberof File
    */
-  public static chmod(filePath: string, permission: number = 0o755): File {
-    fs.chmodSync(filePath, permission);
+  public static chmod(path: string, permission: number = 0o755): File {
+    fs.chmodSync(path, permission);
     return this;
   }
 
   /**
    * Create a temporary directory.
-   * Returns the path to the temporary directory created.
+   *
+   * @static
+   * @return {string} 
+   * @memberof File
    */
   public static makeTmpDirectory(): string {
     const tmp = `${this.getTmpDirectory()}/${uniqid()}/`;
@@ -35,6 +50,12 @@ export default class File {
 
   /**
    * Make a directory.
+   *
+   * @static
+   * @param {string} dirPath Directory path.
+   * @param {number} permission Directory permissions. Default is 0o755.
+   * @return {File}
+   * @memberof File
    */
   public static makeDirectory(dirPath: string, permission: number = 0o755): File {
     if (this.existsFile(dirPath))
@@ -45,11 +66,16 @@ export default class File {
   }
 
   /**
-   * Check if the file exists.
+   * Check if the file or directory exists.
+   *
+   * @static
+   * @param {string} path File or directory path.
+   * @return {boolean} True if the file or directory exists.
+   * @memberof File
    */
-  public static existsFile(filePath: string): boolean {
+  public static existsFile(path: string): boolean {
     try {
-      fs.accessSync(filePath, fs.constants.R_OK | fs.constants.W_OK);
+      fs.accessSync(path, fs.constants.R_OK | fs.constants.W_OK);
       return true;
     } catch {
       return false;
@@ -58,6 +84,10 @@ export default class File {
 
   /**
    * Delete the file.
+   *
+   * @static
+   * @param {string} filePath File Path.
+   * @memberof File
    */
   public static deleteFile(filePath: string): void {
     if (!this.existsFile(filePath))
@@ -67,6 +97,10 @@ export default class File {
 
   /**
    * Delete the directory.
+   *
+   * @static
+   * @param {string} dirPath Directory path.
+   * @memberof File
    */
   public static deleteDirectory(dirPath: string): void {
     fse.removeSync(dirPath);
@@ -74,6 +108,14 @@ export default class File {
 
   /**
    * Write a file
+   *
+   * @static
+   * @param {string} filePath File Path.
+   * @param {string} content The contents of the file. Default is an empty string.
+   * @param {fs.BaseEncodingOptions|string|undefined} options Writing options. Default is undefined.
+   * @param {number} permission File permissions. Default is 0o755.
+   * @return {File}
+   * @memberof File
    */
   public static write(
     filePath: string,
@@ -97,20 +139,35 @@ export default class File {
 
   /**
    * Get the contents of a file as a string.
+   *
+   * @static
+   * @param {string} filePath File Path.
+   * @return {string} The contents of the file.
+   * @memberof File
    */
   public static readAsString(filePath: string): string {
     return fs.readFileSync(filePath).toString();
   }
 
   /**
-   * Obtain the contents of a JSON file as an object.
+   * Get the contents of a JSON file as an object.
+   * 
+   * @static
+   * @param {string} filePath File Path.
+   * @return {object} An object generated from JSON.
+   * @memberof File
    */
   public static readAsJson(filePath: string): {} {
     return JSON.parse(this.readAsString(filePath));
   }
 
   /**
-   * Obtain the contents of the media file as a data URL string.
+   * Get the contents of the media file as a data URL string.
+   *
+   * @static
+   * @param {string} filePath File Path.
+   * @return {string} Data URL.
+   * @memberof File
    */
   public static readAsDataUrl(filePath: string): string {
     const content =  fs.readFileSync(filePath);
@@ -125,14 +182,24 @@ export default class File {
   }
 
   /**
-   * Obtain the contents of a media file as a base64 string.
+   * Get the contents of a media file as a base64 string.
+   *
+   * @static
+   * @param {string} filePath File Path.
+   * @return {string} Base 64 strings.
+   * @memberof File
    */
   public static readAsBase64(filePath: string): string {
     return fs.readFileSync(filePath, {encoding: 'base64'});
   }
 
   /**
-   * Obtain file information.
+   * Get file information.
+   *
+   * @static
+   * @param {string} filePath File Path.
+   * @return {object} File information object.
+   * @memberof File
    */
   public static getStat(filePath: string): any {
     return fs.statSync(filePath);
@@ -140,6 +207,11 @@ export default class File {
 
   /**
    * Get file modification time in unix time.
+   *
+   * @static
+   * @param {string} filePath File Path.
+   * @return {number} Last modified unix time of the file.
+   * @memberof File
    */
   public static getFilemtime(filePath: string): number {
     return moment(this.getStat(filePath).mtime).unix();
@@ -147,6 +219,11 @@ export default class File {
 
   /**
    * Get the file extension.
+   *
+   * @static
+   * @param {string} filePath File Path.
+   * @return {string|undefined} File extension.
+   * @memberof File
    */
   public static getExtension(filePath: string): string|undefined {
     const extension = path.extname(filePath);
@@ -156,30 +233,40 @@ export default class File {
   /**
    * Find files that match the file name or path pattern.
    *
+   * @static
+   * @param {string} pattern File pattern to find.
+   * @param {glob.IOptions} options Options to find. Default is undefined
+   * @return {string[]} Absolute path list of files found.
+   * @memberof File
    * @example
-   * import {File} from 'nodejs-shared';
-   * 
    * File.find('**\/*.js');
    * File.find('**\/glo?.js');
    * File.find('**\/*[0-9]*.js');
-   *
-   * @param  {string} pattern
-   * @param  {glob.IOptions} options
-   * @returns {string[]}
    */
   public static find(pattern: string, options: glob.IOptions = {}) {
     return glob.sync(pattern, {nodir: false, ...options});
   }
 
   /**
-   * Returns the path to the new temporary directory. The directory is not created.
+   * Returns the path to the new temporary directory.
+   * The directory is not created.
+   *
+   * @static
+   * @return {string} Temporary directory path.
+   * @memberof File
    */
   public static getTmpDirectory(): string {
     return os.tmpdir();
   }
 
   /**
-   * Returns the path to the new temporary file. No file is created.
+   * Returns the path to the new temporary file.
+   * No file is created.
+   *
+   * @static
+   * @param {string} extension Extension to be given to temporary files. Default is none.
+   * @return {string} Temporary file path.
+   * @memberof File
    */
   public static getTmpPath(extension?: string): string {
     let filePath = `${this.getTmpDirectory()}/${uniqid()}`;
@@ -190,6 +277,11 @@ export default class File {
 
   /**
    * Check if it is a file.
+   *
+   * @static
+   * @param {string} filePath File Path.
+   * @return {boolean} True if the file is a file.
+   * @memberof File
    */
   public static isFile(filePath: string): boolean {
     try {
@@ -202,13 +294,23 @@ export default class File {
 
   /**
    * Rename a file or directory.
+   *
+   * @static
+   * @param {string} fromPath Original file path.
+   * @param {string} toPath The destination file path.
+   * @memberof File
    */
-  public static rename(from: string, to: string): void {
-    fs.renameSync(from, to);
+  public static rename(fromPath: string, toPath: string): void {
+    fs.renameSync(fromPath, toPath);
   }
 
   /**
    * Check if it is a base64 string.
+   *
+   * @static
+   * @param {string} str String.
+   * @return {boolean} True if base64.
+   * @memberof File
    */
   public static isBase64(str: string): boolean {
     return Buffer.from(str, 'base64').toString('base64') === str;