+++ src/Reflect.ts
@@ -4,6 +4,10 @@
 export default class {
   /**
    * Find static methods from the class.
+   *
+   * @static
+   * @param {any} clazz Class.
+   * @return {Set<string>} Static method name list.
    */
   public static getStaticMethods(clazz: any): Set<string> {
     const methods = new Set<string>();
@@ -21,6 +25,10 @@ export default class {
 
   /**
    * Find a method from an instance.
+   *
+   * @static
+   * @param {any} Class instance.
+   * @return {Set<string>} Instance method name list.
    */
   public static getMethods(instance: any): Set<string> {
     const methods = new Set<string>();