+++ src/Regex.ts
@@ -4,9 +4,12 @@
 export default class {
   /**
    * Escapes the `RegExp` special characters "^", "$", "\", ".", "*", "+", "?", "(", ")", "[", "]", "{", "}", "|" in `string`.
+   *
+   * @static
+   * @param {string} str String.
+   * @param {{[key: string]: string}} replace? Custom replacement characters. The default is none (undefined).
+   * @return {string} Escaped string.
    * @example
-   * import {Regex} from 'nodejs-shared';
-   * 
    * // Escapes the `RegExp` special characters "^", "$", "\", ".", "*", "+", "?", "(", ")", "[", "]", "{", "}", "|" in `string`.
    * Regex.escape('https://example.jp/?tag=TypeScript');// 'https://example\\.jp/\\?tag=TypeScript'
    *