+++ src/interfaces/ImageData.ts
@@ -1,0 +1,9 @@
+export default interface ImageData {
+  buffer: Buffer;
+  height: number;
+  offsetX: number;
+  offsetY: number;
+  width: number;
+  x: number;
+  y: number;
+}