+++ types/interfaces/Dim.d.ts
@@ -1,0 +1,6 @@
+export default interface Dim {
+    top: number;
+    right: number;
+    bottom: number;
+    left: number;
+}