+++ types/interfaces/ImageData.d.ts
@@ -1,0 +1,11 @@
+/// <reference types="node" />
+/// <reference types="node" />
+export default interface ImageData {
+    buffer: Buffer;
+    height: number;
+    offsetX: number;
+    offsetY: number;
+    width: number;
+    x: number;
+    y: number;
+}