+++ CHANGELOG.md
@@ -1,12 +1,45 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.0.11] - 2022/12/15
+### Added
+- Added image merge method.
+    Merge vertically:  
+    <img src="screencaps/merge-images-vertically.png" width="300">
+
+    Merge horizontally:  
+    <img src="screencaps/merge-images-horizontally.png" width="300">
+
+    Set 30px margins between images to merge vertically:  
+    <img src="screencaps/merge-images-vertically-with-margins.png" width="300">
+
+    ```js
+    const {Media} = require('nodejs-shared');
+
+    const imagePaths = ['sample1.png', 'sample2.png', 'sample2.png'];
+
+    // Merge vertically.
+    await Media.mergeImages(imagePaths, 'result1.png', {direction: 'vertical'});
+
+    // Merge horizontally.
+    await Media.mergeImages(imagePaths, 'result2.png', {direction: 'horizontal'});
+
+    // Set 30px margins between images to merge vertically.
+    await Media.mergeImages(imagePaths, 'result3.png', {direction: 'vertical', offset: 30});
+
+    // Set 30px margins between images to merge horizontally.
+    await Media.mergeImages(imagePaths, 'result4.png', {direction: 'horizontal', offset: 30});
+    ```
+
+### Changed
+- API documentation has been moved from github.io to README.md. 
+
 ## [1.0.10] - 2022/10/17
 ### Fixed
 - I am sorry. I forgot to build after fixing the code.
 
 ## [1.0.9] - 2022/10/17
-### Fixed
+### Changed
 - Fix typos in the code.
 
 ## [1.0.8] - 2022/10/17
@@ -36,7 +69,7 @@ All notable changes to this project will be documented in this file.
     ```
 
 ## [1.0.6] - 2022/10/11
-### Fixed
+### Changed
 - Fix documentation.
 
 ## [1.0.5] - 2022/10/11
@@ -44,7 +77,7 @@ All notable changes to this project will be documented in this file.
 - Fixed a bug that prevented the writeDataUrlToFile method of the Media class from writing DataURL format SVG to a file.
 
 ## [1.0.4] - 2022/10/11
-### Fixed
+### Changed
 - The writeDataUrlToFile method of the Media class now automatically adds an extension when the output filename does not have one.  
     Also, the return value now returns the file path instead of the Media class.  
     ```js
@@ -58,7 +91,7 @@ All notable changes to this project will be documented in this file.
     console.log(writtenFilePath);
     // =>test2.png
     ```
-
+### Added
 - Added method to get Mime type from DataURL.
     ```js
     const {Media} = require('nodejs-shared');
@@ -78,7 +111,7 @@ All notable changes to this project will be documented in this file.
     ```
 
 ## [1.0.3] - 2022/8/3
-### Fixed
+### Added
 - Added a method to the Media class to obtain byte size from data URL or base64.
     Get the byte size of data URL:
     ```js
@@ -97,6 +130,8 @@ All notable changes to this project will be documented in this file.
     Media.base64ByteSize(base64);
     // =>30141
     ```
+
+### Changed
 - Changed to the correct method name as a function.  
     * Media.writeBase64Image => Media.writeDataUrlToFile
     * Media.isBase64 => Media.isDataUrl
@@ -105,13 +140,13 @@ All notable changes to this project will be documented in this file.
     * File.readAsBase64 => File.readAsDataUrl
 
 ## [1.0.2] - 2021/10/12
-### Fixed
+### Changed
 - Updated dependent package'sharp'from 0.25.4 to 0.29.1.  
     This update statically links sharp's pre-built libvips binaries, eliminating the need to install Phton.  
     Click [here](https://sharp.pixelplumbing.com/changelog) for sharp change log.
 
 ## [1.0.1] - 2020/9/25
-### Fixed
+### Added
 - Added delete directory method to file module.
     ```js
     import { File } from 'nodejs-shared';
@@ -119,7 +154,7 @@ All notable changes to this project will be documented in this file.
     ````
 
 ## [1.0.0] - 2020/7/13
-### Fixed
+### Added
 - First release.
 
 [1.0.1]: https://github.com/takuya-motoshima/nodejs-shared/compare/v1.0.0...v1.0.1